+++ backend/toy/src/main/java/com/cej/toy/test/controller/TestController.java
@@ -5,16 +5,11 @@ import com.cej.toy.test.domain.dto.LoginResDto;
 import com.cej.toy.test.domain.dto.TestDto;
 import com.cej.toy.test.domain.dto.User;
 import com.cej.toy.test.service.TestService;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.web.authentication.logout.SecurityContextLogoutHandler;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
