+++ backend/toy/src/main/java/com/cej/toy/test/exception/ExceptionHandlers.java
@@ -1,0 +1,8 @@
+package com.cej.toy.test.exception;
+
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+public class ExceptionHandlers {
+
+}
