+++ backend/toy/src/main/java/com/cej/toy/test/service/TestService.java
@@ -8,7 +8,6 @@ import com.cej.toy.test.repository.TestRepository;
 import io.jsonwebtoken.Header;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
-import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
